[20] - Refactorizar código en base a comentarios de PR
