+++ src/app.ts
@@ -4,7 +4,7 @@ import getAllCharger from './getAllCharger';
 import getCharger from './getCharger';
 import getChargesHistory from './getChargesHistory';
 import { getCompanies } from './getCompanies';
-import { insertTokenHandler } from './insertTokenHandler';
+import { insertToken } from './insertToken';
 
 async function app(fastify: FastifyInstance, opts: any){
   fastify.get('/', getMessage);
@@ -12,7 +12,7 @@ async function app(fastify: FastifyInstance, opts: any){
   fastify.get('/charger/:id', getCharger);
   fastify.get('/charges', getChargesHistory);
   fastify.get('/companies', getCompanies);
-  fastify.post('/oauth/token', insertTokenHandler);
+  fastify.post('/oauth/token', insertToken);
 }
 
 export default app;
