+++ src/models/companyModel.ts
@@ -6,10 +6,3 @@ export interface Company {
     employees_count: number;
     secret: string;
 }
-
-export interface Token {
-    access_token: string;
-    expires_in: Date;
-    refresh_token: string;
-    company_id: number;
-}
