+++ src/models/tokenModel.ts
@@ -1,0 +1,6 @@
+export interface Token {
+    access_token: string;
+    expires_in: Date;
+    refresh_token: string;
+    company_id: number;
+}
